refactor(react-gui): split the coloring pane by what each piece decides
ColorPane was 1,535 lines: a dropdown table, a target-key encoding, a
renderer selector, an editable paint table with its own selection and
context menu, six per-mode decks, twelve service-writing callbacks, and
the render that arranges them. Reading any one of those meant scrolling
past the rest, and the file was the largest in the renderer.

The seam used is what a piece decides:

  targetKey.ts          a target is a (kind, uid) pair; this is the encoding
  coloringModes.ts      which items the dropdown offers, as a table
  RendererSelector.tsx  which renderer the coloring applies to
  PaintTable.tsx        how a row is selected, resized, right-clicked
  decks.tsx             what one coloring mode looks like -- six of them,
                        all presentational, which is why the pane can pick
                        one by class name and hand each the same shape
  usePaintSelection.ts  the anchor / selection pair, and why there are two
  useColorPaneActions.ts every write: resolve the target, fire one service

What is left in the pane (503 lines) is the part that is actually about
this pane: reading the coloring state, deciding which deck applies, and
laying it out.

`setCanPastePaint` is now passed into the actions hook rather than
captured, which is what the extraction turned up: Cut / Copy learn the
clipboard state from the write they just did and Paste from finding it
empty, and neither was in a dependency array. Both are now, along with
the pane's own `setSelectedRow` -- three exhaustive-deps warnings that
the split made visible and that are gone rather than suppressed. The
warning baseline drops 75 -> 72.

No markup changed: every extracted block moved verbatim, and the two
local field rows (`LabeledNumberField` / `EnumField`) stay private to the
decks. Promoting them into the form catalog is a catalog decision and
belongs with the other composites.
