+++ tritium/react-gui/src/renderer/components/panes/ColorPane.tsx
@@ -29,124 +29,43 @@
  * This pane is one of the components within the ExplorerView.
  */
 
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
-import {
-    Button,
-    ButtonGroup,
-    Menu,
-    MenuItem,
-    Popover,
-    HTMLSelect,
-    Tooltip,
-} from '@blueprintjs/core'
+import React, { useCallback, useEffect, useMemo, useState } from 'react'
+import { Button, Menu, MenuItem, Popover } from '@blueprintjs/core'
 import { PaneSectionHeader } from './PaneSectionHeader'
-import { AppIcon } from '@renderer/h3-kit/primitives'
-import {
-    Field,
-    FieldGrid,
-    FieldGridRow,
-    ColorField,
-    SelectField,
-    SliderField,
-    RejectNumberInput,
-} from '../../h3-kit/form'
+import { Field, SelectField } from '../../h3-kit/form'
 import type { RendColoringId } from '@shared/types/sceneCtxMenu'
-import type {
-    BfacParams,
-    ColoringTargetKind,
-    CpkColors,
-    ElePotMapObjectEntry,
-    ElepotParams,
-    PaintCapableRendererEntry,
-    PaintEntryDto,
-    RainbowParams,
-} from '../../worker/server/services/rendererColoring.service'
+import type { ColoringTargetKind } from '../../worker/server/services/rendererColoring.service'
 import type { SceneObjectEntry } from '../../worker/server/services/listSceneObjects.service'
-import { CueColorField, ColorPickerProvider } from '@renderer/h3-kit/colorpicker'
-import { scrollRowIntoView, useListKeyNav } from '../../h3-kit/list'
+import { ColorPickerProvider } from '@renderer/h3-kit/colorpicker'
 import { MultiGradSection } from '../multigrad/MultiGradSection'
 import { usePaintCapableRenderers } from '../../hooks/usePaintCapableRenderers'
 import { usePaintColoringStyles } from '../../hooks/usePaintColoringStyles'
 import { useRendererColoringState } from '../../hooks/useRendererColoringState'
 import { useElePotMapObjects } from '../../hooks/useElePotMapObjects'
 import { useMolCoordObjects } from '../../hooks/useMolCoordObjects'
-import { PaintSelCell } from './PaintSelCell'
-import { fireService } from '../../utils/fireService'
 import { IPC } from '@shared/ipcChannels'
 import { useClipboardScope } from '../../hooks/useClipboardScope'
-import { useColumnResize } from '@renderer/hooks/useColumnResize'
-import { useShowContextMenu } from '../menu/ContextMenuProvider'
-import type { MenuNode } from '@shared/menuNodes'
 import { useCueMol } from '../../hooks/cuemol/useCueMol'
 import { useActiveScene } from '../../state/workspace'
-
-// ------------------------------------------------------------
-// Coloring type dropdown items
-//
-// The deck colour editors below use the reusable `CueColorField`, which
-// reads `cm` / `sceneId` from the `ColorPickerProvider` wrapped around this
-// pane's body (so they need no prop threading).
-// ------------------------------------------------------------
-
-interface ColoringModeItem {
-    label: string
-    coloringId: RendColoringId | null
-    enabled: boolean
-    /**
-     * UXP `setupColoringSelector` hides the Electrostatic-potential item
-     * unless the renderer is `molsurf` / `dsurface`. We mirror that with a
-     * per-item gate that the parent component evaluates against
-     * `state.surfaceType`.
-     */
-    surfaceOnly?: boolean
-    /** Shown only when the renderer exposes a `multi_grad` gradient. */
-    multigradOnly?: boolean
-}
-
-/**
- * Marker for the "Paint coloring" row: it is not a leaf item but a submenu
- * whose entries are built at render time (see `paintSubmenuItems`).
- */
-const PAINT_SUBMENU_ID = 'paint-type-paint'
-
-/**
- * Default width of the paint table's Selection column, in px; Color takes
- * whatever is left. Persisted per user so the split survives a restart --
- * UXP persisted the treecol widths (`persist="hidden ordinal width"`) for
- * the same reason.
- */
-const PAINT_COL_WIDTHS = { selection: 130 }
-const PAINT_COL_WIDTHS_KEY = 'cuemol.colorPane.paint.colWidths'
-
-/**
- * Width the Color column keeps no matter how far the splitter is dragged.
- * The table is `table-layout: fixed`, so an oversized Selection column would
- * otherwise push the table past its wrapper -- which clips the selected
- * row's outline and can hide the colour control entirely.
- */
-const PAINT_MIN_COLOR_COL = 90
-
-/** Floor for the Selection column, matching useColumnResize's own minimum. */
-const PAINT_COL_MIN = 40
-
-/** Actions the paint row context menu can return. */
-type PaintCtxAction = 'cut' | 'copy' | 'paste' | 'delete' | 'deleteAll'
-
-const COLORING_MODE_ITEMS: ColoringModeItem[] = [
-    { label: 'Paint coloring',          coloringId: 'paint-type-paint',    enabled: true  },
-    { label: 'Solid coloring',          coloringId: 'paint-type-solid',    enabled: true  },
-    // The three CPK entries differ only in the carbon colour, matching the
-    // Default / Darkgray / Lightgray CPK styles the renderer context menu
-    // offers. Labels follow those styles' `desc`.
-    { label: 'CPK coloring',            coloringId: 'paint-type-cpk',      enabled: true  },
-    { label: 'CPK (darkgray carbon)',   coloringId: 'paint-type-cpk-darkgray',  enabled: true },
-    { label: 'CPK (lightgray carbon)',  coloringId: 'paint-type-cpk-lightgray', enabled: true },
-    { label: 'Bfac/Occ coloring',       coloringId: 'paint-type-bfac',     enabled: true  },
-    { label: 'Rainbow coloring',        coloringId: 'paint-type-rainbow',  enabled: true  },
-    { label: 'Electrostatic potential', coloringId: 'paint-type-elepot',   enabled: true, surfaceOnly: true },
-    { label: 'Multi-gradient coloring', coloringId: 'paint-type-multigrad', enabled: true, multigradOnly: true },
-    { label: 'Reset to default style',  coloringId: 'paint-type-resetdef', enabled: true  },
-]
+import {
+    COLORING_MODE_ITEMS,
+    PAINT_DECK_CLASS,
+    PAINT_SUBMENU_ID,
+    SOLID_DECK_CLASSES,
+} from './colorPane/coloringModes'
+import { makeKey, parseTargetKey, type TargetKey } from './colorPane/targetKey'
+import { RendererSelector } from './colorPane/RendererSelector'
+import { usePaintSelection } from './colorPane/usePaintSelection'
+import { useColorPaneActions } from './colorPane/useColorPaneActions'
+import { PaintTable } from './colorPane/PaintTable'
+import {
+    BfacDeck,
+    CpkDeck,
+    DeferredDeck,
+    ElepotDeck,
+    RainbowDeck,
+    SolidDeck,
+} from './colorPane/decks'
 
 // ------------------------------------------------------------
 // Component props
@@ -158,720 +77,9 @@ interface ColorPaneProps {
 }
 
 // ------------------------------------------------------------
-// Sub-rendering helpers
-// ------------------------------------------------------------
-
-/**
- * Encode a target row as a string the `<select>` element can carry. Format:
- * `"<kind>:<uid>"`. `parseTargetKey` is the inverse.
- */
-type TargetKey = string
-const makeKey = (kind: ColoringTargetKind, id: number): TargetKey =>
-    `${kind}:${id}`
-function parseTargetKey(
-    key: string,
-): { targetKind: ColoringTargetKind; id: number } | null {
-    const sep = key.indexOf(':')
-    if (sep < 0) return null
-    const kind = key.slice(0, sep)
-    const id = Number(key.slice(sep + 1))
-    if (kind !== 'object' && kind !== 'renderer') return null
-    if (Number.isNaN(id)) return null
-    return { targetKind: kind, id }
-}
-
-interface RendererSelectorProps {
-    renderers: PaintCapableRendererEntry[]
-    selectedKey: TargetKey | null
-    onChange: (key: TargetKey | null) => void
-}
-
-const RendererSelector: React.FC<RendererSelectorProps> = ({
-    renderers,
-    selectedKey,
-    onChange,
-}) => {
-    // Group by parent object so the object row sits at the top of its
-    // <optgroup>, followed by its child renderers (mirrors UXP layout).
-    const groups = useMemo(() => {
-        const byObj = new Map<number, { objName: string; entries: PaintCapableRendererEntry[] }>()
-        for (const r of renderers) {
-            const g = byObj.get(r.objId) ?? { objName: r.objName, entries: [] }
-            g.entries.push(r)
-            byObj.set(r.objId, g)
-        }
-        return Array.from(byObj.values())
-    }, [renderers])
-
-    if (renderers.length === 0) {
-        return (
-            <HTMLSelect
-                disabled
-                fill
-                value=""
-                onChange={() => {}}
-                className="color-enum-select h3-form-select"
-            >
-                <option value="">(no paint-capable renderers)</option>
-            </HTMLSelect>
-        )
-    }
-
-    return (
-        <HTMLSelect
-            fill
-            value={selectedKey ?? ''}
-            className="color-enum-select h3-form-select"
-            onChange={(e) => {
-                const v = e.target.value
-                onChange(v === '' ? null : v)
-            }}
-        >
-            {groups.map((g, gi) => (
-                <optgroup key={gi} label={g.objName || '(unnamed)'}>
-                    {g.entries.map((r) => {
-                        const key = makeKey(r.targetKind, r.rendId)
-                        const label =
-                            r.targetKind === 'object'
-                                ? `${r.name || r.typeName} (object)`
-                                : `${r.name || r.typeName} (${r.typeName})`
-                        return (
-                            <option key={key} value={key}>
-                                {label}
-                            </option>
-                        )
-                    })}
-                </optgroup>
-            ))}
-        </HTMLSelect>
-    )
-}
-
-interface PaintTableProps {
-    entries: PaintEntryDto[]
-    /** Anchor row: drives Add's insert point and Move up/down. */
-    selectedIdx: number | null
-    /** Whole selection; the multi-target actions read this. */
-    selectedIdxs: Set<number>
-    onSelect: (idx: number) => void
-    /** Cmd/Ctrl+click toggle. */
-    onToggleSelect: (idx: number) => void
-    /** Shift+click range from the anchor; `additive` unions instead of replacing. */
-    onSelectRange: (idx: number, additive: boolean) => void
-    onAdd: () => void
-    onRemove: () => void
-    onMoveUp: () => void
-    onMoveDown: () => void
-    onUpdate: (idx: number, field: 'selStr' | 'colorValue', value: string) => void
-    /** Clear the whole list (UXP `paintpanel-delallbtn`). */
-    onRemoveAll: () => void
-    onCut: () => void
-    onCopy: () => void
-    onPaste: () => void
-    /** True while the worker-local paint clipboard holds at least one row. */
-    canPaste: boolean
-    /** sceneId required for MolSelList named-def lookup. */
-    sceneId: number
-    /**
-     * Parent mol uid (for renderer-row targets, the renderer's parent
-     * object; for object-row targets, the object itself). Forwarded to
-     * MolSelList so the picker shows the molecule's "current (<sel>)"
-     * preset and any mol-scope named defs.
-     */
-    molId?: number
-}
-
-const PaintTable: React.FC<PaintTableProps> = ({
-    entries,
-    selectedIdx,
-    selectedIdxs,
-    onSelect,
-    onToggleSelect,
-    onSelectRange,
-    onAdd,
-    onRemove,
-    onMoveUp,
-    onMoveDown,
-    onUpdate,
-    onRemoveAll,
-    onCut,
-    onCopy,
-    onPaste,
-    canPaste,
-    sceneId,
-    molId,
-}) => {
-    /** Row ids in display order, for the shared keyboard navigation. */
-    const rowIds = useMemo(() => entries.map((e) => String(e.idx)), [entries])
-
-    const isRowSelected = selectedIdxs.size > 0
-    const isSingleRow = selectedIdxs.size === 1
-    const showContextMenu = useShowContextMenu()
-
-    // Drag-resizable split between the Selection and Color columns, the
-    // tritium home of UXP's `<splitter class="tree-splitter"/>` between the
-    // `paint_name` and `paint_value` treecols. Only the first column is
-    // sized; Color absorbs the remainder, as in GenericTab.
-    const { widths, startResize } = useColumnResize(
-        PAINT_COL_WIDTHS,
-        undefined,
-        PAINT_COL_WIDTHS_KEY,
-    )
-
-    // Measure the table wrapper so the stored Selection width can be clamped
-    // against the space actually available. The stored value is left alone:
-    // widening the panel again restores the width the user chose.
-    const wrapRef = useRef<HTMLDivElement | null>(null)
-    const [wrapWidth, setWrapWidth] = useState(0)
-    useEffect(() => {
-        const el = wrapRef.current
-        if (!el || typeof ResizeObserver === 'undefined') return
-        const ro = new ResizeObserver(() => setWrapWidth(el.clientWidth))
-        ro.observe(el)
-        setWrapWidth(el.clientWidth)
-        return () => ro.disconnect()
-    }, [])
-    const selWidth =
-        wrapWidth > 0
-            ? Math.max(
-                  PAINT_COL_MIN,
-                  Math.min(widths.selection, wrapWidth - PAINT_MIN_COLOR_COL),
-              )
-            : widths.selection
-
-    /**
-     * A modifier-click is a row-selection gesture, not a text gesture: the
-     * browser's default for Shift+mousedown is to extend the DOM text
-     * selection from the last caret, which painted every row between the
-     * anchor and the click as selected text. Cancelling the default here
-     * leaves the click itself (and `onRowClick` below) intact.
-     */
-    const onRowMouseDown = useCallback((e: React.MouseEvent) => {
-        if (e.shiftKey || e.metaKey || e.ctrlKey) e.preventDefault()
-    }, [])
-
-    /**
-     * Row click with the standard modifiers: Cmd/Ctrl toggles, Shift ranges
-     * from the anchor (Shift+Cmd unions), a plain click replaces.
-     */
-    const onRowClick = useCallback(
-        (idx: number, e: React.MouseEvent) => {
-            if (e.shiftKey) {
-                onSelectRange(idx, e.metaKey || e.ctrlKey)
-                return
-            }
-            if (e.metaKey || e.ctrlKey) {
-                onToggleSelect(idx)
-                return
-            }
-            onSelect(idx)
-        },
-        [onSelect, onToggleSelect, onSelectRange],
-    )
-
-    /**
-     * Right-click menu: the clipboard trio plus the destructive actions,
-     * matching UXP's `paintPanelCtxtMenu` -- Cut / Copy / Paste lived there,
-     * never on the toolbar. The accelerators shown here are the ones the app
-     * menu already owns; they reach this deck through the `paint-deck`
-     * clipboard scope, so the menu is a discoverability surface for keys that
-     * work with or without it.
-     *
-     * Right-clicking a row that is already part of a multi-selection keeps
-     * that selection, so the menu acts on all of it; right-clicking elsewhere
-     * selects just that row first (scene-tree parity).
-     */
-    const onRowContextMenu = useCallback(
-        (idx: number | null, e: React.MouseEvent) => {
-            e.preventDefault()
-            if (idx !== null && !selectedIdxs.has(idx)) onSelect(idx)
-            const rows = idx !== null || isRowSelected
-            const nodes: MenuNode<PaintCtxAction>[] = [
-                { label: 'Cut', accelerator: 'CmdOrCtrl+X', enabled: rows, action: 'cut' },
-                { label: 'Copy', accelerator: 'CmdOrCtrl+C', enabled: rows, action: 'copy' },
-                { label: 'Paste', accelerator: 'CmdOrCtrl+V', enabled: canPaste, action: 'paste' },
-                { type: 'separator' },
-                { label: 'Delete', enabled: rows, action: 'delete' },
-                { label: 'Delete all', enabled: entries.length > 0, action: 'deleteAll' },
-            ]
-            void showContextMenu(nodes, { x: e.clientX, y: e.clientY }).then(
-                (action) => {
-                    switch (action) {
-                        case 'cut': onCut(); break
-                        case 'copy': onCopy(); break
-                        case 'paste': onPaste(); break
-                        case 'delete': onRemove(); break
-                        case 'deleteAll': onRemoveAll(); break
-                        default: break
-                    }
-                },
-            )
-        },
-        [
-            showContextMenu, selectedIdxs, isRowSelected, onSelect, entries.length,
-            onCut, onCopy, onPaste, onRemove, onRemoveAll, canPaste,
-        ],
-    )
-
-    /**
-     * Arrow / Home / End over the rows, the same binding the scene tree and
-     * every other list uses (h3-kit/list). Rows carry `data-row-idx` so the
-     * moved-to row can be scrolled into view.
-     */
-    const navKeyDown = useListKeyNav({
-        items: rowIds,
-        activeId: selectedIdx === null ? null : String(selectedIdx),
-        onSelect: (id) => onSelect(Number(id)),
-        onSelectRange: (id, _items, additive) => onSelectRange(Number(id), additive),
-        onScrollTo: (id) => {
-            scrollRowIntoView(wrapRef.current, `[data-row-idx="${id}"]`)
-        },
-    })
-
-    return (
-        <>
-            <div className="color-section-label">Paint coloring:</div>
-            {/* Marks the paint deck as the target of Edit > Cut/Copy/Paste
-                while the user is working here. tabIndex keeps the wrapper
-                focusable so a row click parks focus inside the scope; the
-                handlers are registered by useClipboardScope below. */}
-            <div
-                ref={wrapRef}
-                className="color-table-wrap"
-                tabIndex={-1}
-                data-clipboard-scope="paint-deck"
-                onKeyDown={navKeyDown}
-                style={{ outline: 'none' }}
-            >
-                <table className="color-table">
-                    <colgroup>
-                        <col style={{ width: selWidth }} />
-                        {/* Color takes the remaining width */}
-                        <col />
-                    </colgroup>
-                    <thead>
-                        <tr>
-                            <th className="color-th-selection">
-                                <span className="color-th-label">Selection</span>
-                                <div
-                                    className="color-resize-handle"
-                                    onMouseDown={(e) => startResize('selection', e)}
-                                />
-                            </th>
-                            <th className="color-th-color">Color</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        {entries.length === 0 ? (
-                            <tr onContextMenu={(e) => onRowContextMenu(null, e)}>
-                                <td colSpan={2} className="color-empty-row">
-                                    (no paint entries — click + to add)
-                                </td>
-                            </tr>
-                        ) : (
-                            entries.map((entry) => (
-                                <tr
-                                    key={entry.idx}
-                                    data-row-idx={entry.idx}
-                                    className={`color-row ${selectedIdxs.has(entry.idx) ? 'selected' : ''}`}
-                                    onMouseDown={onRowMouseDown}
-                                    onClick={(e) => onRowClick(entry.idx, e)}
-                                    onContextMenu={(e) =>
-                                        onRowContextMenu(entry.idx, e)
-                                    }
-                                >
-                                    <td className="color-cell-selection">
-                                        <PaintSelCell
-                                            sceneID={sceneId}
-                                            molID={molId}
-                                            value={entry.selStr}
-                                            onFocus={() => onSelect(entry.idx)}
-                                            onCommit={(v) =>
-                                                onUpdate(entry.idx, 'selStr', v)
-                                            }
-                                        />
-                                    </td>
-                                    <td className="color-cell-color">
-                                        <CueColorField
-                                            value={entry.colorValue ?? ''}
-                                            onCommit={(v) =>
-                                                onUpdate(entry.idx, 'colorValue', v)
-                                            }
-                                        />
-                                    </td>
-                                </tr>
-                            ))
-                        )}
-                    </tbody>
-                </table>
-            </div>
-
-            <div className="color-actions" data-clipboard-scope="paint-deck">
-                <ButtonGroup minimal>
-                    <Tooltip content="Add row" placement="top" compact>
-                        <Button
-                            small
-                            icon={<AppIcon name="ui.add" aria-hidden />}
-                            aria-label="Add row"
-                            className="color-action-btn"
-                            onClick={onAdd}
-                        />
-                    </Tooltip>
-                    <Tooltip content="Remove row" placement="top" compact>
-                        <Button
-                            small
-                            icon={<AppIcon name="ui.remove" aria-hidden />}
-                            aria-label="Remove row"
-                            className="color-action-btn"
-                            onClick={onRemove}
-                            disabled={!isRowSelected}
-                        />
-                    </Tooltip>
-                    <Tooltip content="Move up" placement="top" compact>
-                        <Button
-                            small
-                            icon={<AppIcon name="ui.arrowUp" aria-hidden />}
-                            aria-label="Move row up"
-                            className="color-action-btn"
-                            onClick={onMoveUp}
-                            disabled={!isSingleRow || selectedIdx === 0}
-                        />
-                    </Tooltip>
-                    <Tooltip content="Move down" placement="top" compact>
-                        <Button
-                            small
-                            icon={<AppIcon name="ui.arrowDown" aria-hidden />}
-                            aria-label="Move row down"
-                            className="color-action-btn"
-                            onClick={onMoveDown}
-                            disabled={
-                                !isSingleRow ||
-                                (selectedIdx !== null && selectedIdx >= entries.length - 1)
-                            }
-                        />
-                    </Tooltip>
-                </ButtonGroup>
-            </div>
-        </>
-    )
-}
-
-interface SolidDeckProps {
-    /** UXP-style coloring class name; empty string when coloring is null. */
-    className: string
-    defaultColor: string
-    onCommit: (color: string) => void
-}
-
-const SolidDeck: React.FC<SolidDeckProps> = ({ className, defaultColor, onCommit }) => (
-    <div className="color-solid-deck">
-        <div className="color-section-label">
-            {className === '' ? 'Solid coloring' : className}
-        </div>
-        <Field label="Default color" inline>
-            <ColorField value={defaultColor} onCommit={onCommit} />
-        </Field>
-    </div>
-)
-
-interface DeferredDeckProps {
-    className: string
-}
-
-/**
- * Fallback deck for any coloring class `renderClassDeck` has no editor
- * for. It is reachable in normal use, not dead code: a scene can carry
- * coloring classes the panel does not edit (`ScriptColoring`, for one),
- * so the deck names the class and points at the modes that are editable
- * rather than rendering an empty panel.
- */
-const DeferredDeck: React.FC<DeferredDeckProps> = ({ className }) => (
-    <div className="color-deferred-deck">
-        <div className="color-section-label">{className}</div>
-        <p className="color-deferred-note">
-            Editing this coloring mode is not yet implemented. Switch to
-            Paint or Solid via the dropdown, or use Reset to default style.
-        </p>
-    </div>
-)
-
-interface CpkDeckProps {
-    colors: CpkColors
-    onCommit: (propName: string, value: string) => void
-}
-
-/** The deck's rows, in UXP `coloring-deck-cpk.xul` order. */
-const CPK_ELEMENTS: { label: string; prop: string; key: keyof CpkColors }[] = [
-    { label: 'Carbon',     prop: 'col_C', key: 'colC' },
-    { label: 'Nitrogen',   prop: 'col_N', key: 'colN' },
-    { label: 'Oxygen',     prop: 'col_O', key: 'colO' },
-    { label: 'Sulfur',     prop: 'col_S', key: 'colS' },
-    { label: 'Phosphorus', prop: 'col_P', key: 'colP' },
-    { label: 'Hydrogen',   prop: 'col_H', key: 'colH' },
-    { label: 'Others',     prop: 'col_X', key: 'colX' },
-]
-
-/**
- * Mirrors UXP `coloring-deck-cpk.xul`: 7 per-element colour pickers.
- *
- * Laid out with `FieldGrid` rather than a stack of inline `Field`s so the
- * labels share one column. With per-row labels each one ellipsised at its own
- * width once the pane narrowed ("Phosp..." vs "Sul..."), which pushed every
- * swatch to a different x -- the element names here are long enough and
- * different enough in length for that to show at ordinary pane widths.
- */
-const CpkDeck: React.FC<CpkDeckProps> = ({ colors, onCommit }) => (
-    <div className="color-deck-scroll">
-        <div className="color-section-label">CPK coloring:</div>
-        <FieldGrid>
-            {CPK_ELEMENTS.map(({ label, prop, key }) => (
-                <FieldGridRow key={prop} label={label}>
-                    <ColorField
-                        value={colors[key]}
-                        onCommit={(v) => onCommit(prop, v)}
-                    />
-                </FieldGridRow>
-            ))}
-        </FieldGrid>
-    </div>
-)
-
-/**
- * Bfac / elepot numeric field with a label column. Wraps the catalog
- * `RejectNumberInput` (reject-and-revert validation -- out-of-range / NaN is
- * dropped silently, matching UXP `onRainbowChange` / `onBfacChange`) in a
- * `Field` so the label and reject input compose like every other form row.
- */
-interface LabeledNumberFieldProps {
-    label: string
-    value: number
-    min?: number
-    max?: number
-    scale?: number
-    decimals?: number
-    onCommit: (next: number) => void
-    disabled?: boolean
-}
-
-const LabeledNumberField: React.FC<LabeledNumberFieldProps> = ({
-    label, value, min, max, scale, decimals, onCommit, disabled,
-}) => (
-    <Field label={label} inline>
-        <RejectNumberInput
-            value={value}
-            min={min}
-            max={max}
-            scale={scale}
-            decimals={decimals}
-            onCommit={onCommit}
-            disabled={disabled}
-        />
-    </Field>
-)
-
-/**
- * Shared select with a label column: a catalog `SelectField` inside a `Field`
- * so Bfac / Rainbow mode dropdowns compose like every other labeled row.
- */
-interface EnumFieldProps {
-    label: string
-    value: string
-    options: { value: string; label: string }[]
-    onCommit: (next: string) => void
-    disabled?: boolean
-}
-
-const EnumField: React.FC<EnumFieldProps> = ({
-    label, value, options, onCommit, disabled,
-}) => (
-    <Field label={label} inline>
-        <SelectField value={value} disabled={disabled} onChange={onCommit}>
-            {options.map((o) => (
-                <option key={o.value} value={o.value}>
-                    {o.label}
-                </option>
-            ))}
-        </SelectField>
-    </Field>
-)
-
-interface RainbowDeckProps {
-    params: RainbowParams
-    onCommit: (propName: string, value: string | number) => void
-}
-
-/** Mirrors UXP `coloring-deck-rainbow.xul` -- the four numerics use the
- * `<numslider>` widget (slider + numeric spinbox + unit). */
-const RainbowDeck: React.FC<RainbowDeckProps> = ({ params, onCommit }) => (
-    <div className="color-deck-scroll">
-        <div className="color-section-label">Rainbow coloring:</div>
-        <EnumField
-            label="Mode" value={params.mode}
-            options={[
-                { value: 'mol',   label: 'Molecule' },
-                { value: 'chain', label: 'Chain'    },
-            ]}
-            onCommit={(v) => onCommit('mode', v)}
-        />
-        <EnumField
-            label="Change by" value={params.incrMode}
-            options={[
-                { value: 'chain',  label: 'Chain'        },
-                { value: 'resid',  label: 'Residue'      },
-                { value: 'protss', label: 'Prot secstr'  },
-            ]}
-            onCommit={(v) => onCommit('incr_mode', v)}
-        />
-        <SliderField
-            label="Start H" value={params.startHue} min={0} max={360} unit="°"
-            onCommit={(v) => onCommit('start_hue', v)}
-        />
-        <SliderField
-            label="End H" value={params.endHue} min={0} max={360} unit="°"
-            onCommit={(v) => onCommit('end_hue', v)}
-        />
-        <SliderField
-            label="Brightness" value={params.brightness} min={0} max={100} scale={100} unit="%"
-            onCommit={(v) => onCommit('bri', v)}
-        />
-        <SliderField
-            label="Saturation" value={params.saturation} min={0} max={100} scale={100} unit="%"
-            onCommit={(v) => onCommit('sat', v)}
-        />
-    </div>
-)
-
-interface BfacDeckProps {
-    params: BfacParams
-    onCommit: (propName: string, value: string | number) => void
-}
-
-/** Mirrors UXP `coloring-deck-bfac.xul`. */
-const BfacDeck: React.FC<BfacDeckProps> = ({ params, onCommit }) => {
-    const manual = params.autoMode === 'none'
-    return (
-        <div className="color-deck-scroll">
-            <div className="color-section-label">Bfac coloring:</div>
-            <EnumField
-                label="Mode" value={params.mode}
-                options={[
-                    { value: 'bfac',   label: 'B-factor'              },
-                    { value: 'occ',    label: 'Occupancy'             },
-                    { value: 'center', label: 'Distance from center'  },
-                ]}
-                onCommit={(v) => onCommit('mode', v)}
-            />
-            <Field label="Low" inline>
-                <ColorField value={params.lowColor} onCommit={(v) => onCommit('lowcol', v)} />
-            </Field>
-            <Field label="High" inline>
-                <ColorField value={params.highColor} onCommit={(v) => onCommit('highcol', v)} />
-            </Field>
-            <div className="color-section-sublabel">Parameter</div>
-            <EnumField
-                label="Auto" value={params.autoMode}
-                options={[
-                    { value: 'none', label: 'Manual'        },
-                    { value: 'mol',  label: 'Auto (by mol)' },
-                    { value: 'rend', label: 'Auto (by rend)' },
-                ]}
-                onCommit={(v) => onCommit('auto', v)}
-            />
-            <LabeledNumberField
-                label="Low" value={params.lowParam} disabled={!manual}
-                onCommit={(v) => onCommit('lowpar', v)}
-            />
-            <LabeledNumberField
-                label="High" value={params.highParam} disabled={!manual}
-                onCommit={(v) => onCommit('highpar', v)}
-            />
-        </div>
-    )
-}
-
-interface ElepotDeckProps {
-    params: ElepotParams
-    /** ElePotMap objects available for the "potential object" selector. */
-    objects: ElePotMapObjectEntry[]
-    onCommit: (propName: string, value: string | number | boolean) => void
-}
-
-/**
- * Mirrors UXP `coloring-deck-elepot.xul`. Properties live on the surface
- * renderer itself (not on a ColoringScheme); the deck appears whenever the
- * renderer's `colormode === "potential"`. The ramp rows are a genuine
- * 3-column layout (label, numeric param, colour swatch); everything else
- * reuses the catalog SelectField / ColorField / RejectNumberInput.
- */
-const ElepotDeck: React.FC<ElepotDeckProps> = ({ params, objects, onCommit }) => (
-    <div className="color-deck-scroll">
-        <div className="color-section-label">Elepot coloring:</div>
-        <Field label="Potential" inline>
-            <SelectField
-                value={params.elepot}
-                disabled={objects.length === 0}
-                onChange={(v) => onCommit('elepot', v)}
-            >
-                {/* When the renderer's elepot is unset or points to a now-deleted
-                  * object, show a sentinel row so the dropdown is still
-                  * controlled. UXP shows the same "(none)" state via
-                  * `mPotSel.selectObjectByName("")`. */}
-                {objects.find((o) => o.name === params.elepot) === undefined && (
-                    <option value={params.elepot}>
-                        {params.elepot || '(no ElePotMap selected)'}
-                    </option>
-                )}
-                {objects.map((o) => (
-                    <option key={o.objId} value={o.name}>
-                        {o.name}
-                    </option>
-                ))}
-            </SelectField>
-        </Field>
-        <div className="color-field-row">
-            <label className="color-field-label">By SAS</label>
-            <input
-                type="checkbox"
-                checked={params.rampAbove}
-                onChange={(e) => onCommit('ramp_above', e.target.checked)}
-            />
-        </div>
-        <div className="color-field-row color-elepot-ramp-row">
-            <label className="color-elepot-ramp-label">High</label>
-            <RejectNumberInput
-                className="color-elepot-number"
-                value={params.highParam}
-                onCommit={(v) => onCommit('highpar', v)}
-            />
-            <ColorField value={params.highColor} onCommit={(v) => onCommit('highcol', v)} />
-        </div>
-        <div className="color-field-row color-elepot-ramp-row">
-            <label className="color-elepot-ramp-label">Mid</label>
-            <RejectNumberInput
-                className="color-elepot-number"
-                value={params.midParam}
-                onCommit={(v) => onCommit('midpar', v)}
-            />
-            <ColorField value={params.midColor} onCommit={(v) => onCommit('midcol', v)} />
-        </div>
-        <div className="color-field-row color-elepot-ramp-row">
-            <label className="color-elepot-ramp-label">Low</label>
-            <RejectNumberInput
-                className="color-elepot-number"
-                value={params.lowParam}
-                onCommit={(v) => onCommit('lowpar', v)}
-            />
-            <ColorField value={params.lowColor} onCommit={(v) => onCommit('lowcol', v)} />
-        </div>
-    </div>
-)
-
-// ------------------------------------------------------------
 // Main component
 // ------------------------------------------------------------
 
-const PAINT_DECK_CLASS = 'PaintColoring'
-const SOLID_DECK_CLASSES = new Set(['', 'SolidColoring'])
 
 export const ColorPane: React.FC<ColorPaneProps> = ({ collapsed, onToggleCollapse }) => {
     const { cm } = useCueMol()
@@ -882,52 +90,9 @@ export const ColorPane: React.FC<ColorPaneProps> = ({ collapsed, onToggleCollaps
     // Selected row is stored as a key (`<kind>:<uid>`) so a single state
     // captures both the target kind and the C++ uid. Parsed at use sites.
     const [selectedKey, setSelectedKey] = useState<TargetKey | null>(null)
-    // `selectedRow` is the anchor / primary row: it drives the single-target
-    // operations (Add insert point, Move up/down) and is the origin a
-    // Shift+click ranges from. `selectedRows` is the full selection the
-    // multi-target operations (Delete, Cut, Copy, paste anchor) act on; it
-    // always contains `selectedRow` while anything is selected.
-    const [selectedRow, setSelectedRowState] = useState<number | null>(null)
-    const [selectedRows, setSelectedRows] = useState<Set<number>>(() => new Set())
-
-    /** Replace the selection with a single row (or clear it with null). */
-    const setSelectedRow = useCallback((idx: number | null) => {
-        setSelectedRowState(idx)
-        setSelectedRows(idx === null ? new Set() : new Set([idx]))
-    }, [])
-
-    /** Cmd/Ctrl+click: toggle one row in the selection. */
-    const toggleSelectedRow = useCallback((idx: number) => {
-        setSelectedRows((prev) => {
-            const next = new Set(prev)
-            if (next.has(idx)) next.delete(idx)
-            else next.add(idx)
-            setSelectedRowState(next.size === 0 ? null : idx)
-            return next
-        })
-    }, [])
-
-    /**
-     * Shift+click: select the range between the anchor and `idx`.
-     *
-     * The anchor stays put so repeated Shift+clicks re-extend from the same
-     * origin (Finder parity, matching the scene tree's `selectRangeTo`).
-     */
-    const selectRowRange = useCallback((idx: number, additive: boolean) => {
-        setSelectedRowState((anchor) => {
-            if (anchor === null) {
-                setSelectedRows(new Set([idx]))
-                return idx
-            }
-            const [lo, hi] = anchor <= idx ? [anchor, idx] : [idx, anchor]
-            const range: number[] = []
-            for (let i = lo; i <= hi; ++i) range.push(i)
-            setSelectedRows((prev) =>
-                additive ? new Set([...prev, ...range]) : new Set(range),
-            )
-            return anchor
-        })
-    }, [])
+    const {
+        selectedRow, selectedRows, setSelectedRow, toggleSelectedRow, selectRowRange,
+    } = usePaintSelection()
 
     // Auto-select the first row when the list changes, and clear when the
     // active row disappears.
@@ -949,7 +114,7 @@ export const ColorPane: React.FC<ColorPaneProps> = ({ collapsed, onToggleCollaps
     // Reset paint-row selection when the active target changes.
     useEffect(() => {
         setSelectedRow(null)
-    }, [selectedKey])
+    }, [selectedKey, setSelectedRow])
 
     // Whether the OS clipboard holds paint rows; gates the Paste button.
     // Electron has no clipboard-change event, so this is re-asked on mount
@@ -1071,210 +236,14 @@ export const ColorPane: React.FC<ColorPaneProps> = ({ collapsed, onToggleCollaps
         [cm, sceneId, target],
     )
 
-    const onSelectMode = useCallback(
-        (coloringId: RendColoringId) => {
-            const t = requireTarget()
-            if (!t || !cm) return
-            fireService(cm, 'setRendererColoring', {
-                ...t,
-                coloringId,
-            })
-        },
-        [cm, requireTarget],
-    )
-
-    const onAddRow = useCallback(() => {
-        const t = requireTarget()
-        if (!t || !cm) return
-        // UXP `onAddCmd` parity:
-        //   - insert before the selected row (id = elem.obj_id);
-        //     when no row is selected, id = 0 (insert at top).
-        //   - inherit sel / color from the selected row (UXP also checks
-        //     the parent mol's `sel` first; deferred to a later phase).
-        //   - fall back to "*" / "#FFF" when there is no row to inherit
-        //     from (paint-propdlg defaults).
-        const idx = selectedRow !== null ? selectedRow : 0
-        const ref = selectedRow !== null ? entries[selectedRow] : undefined
-        const selStr = ref?.selStr ?? '*'
-        const colorValue = ref?.colorValue ?? '#FFFFFF'
-        fireService(cm, 'addPaintEntry', {
-            ...t,
-            idx,
-            selStr,
-            colorValue,
-        })
-        // The new entry occupies the insert position; track it so the
-        // toolbar buttons act on the freshly-added row.
-        setSelectedRow(idx)
-    }, [cm, requireTarget, selectedRow, entries, setSelectedRow])
-
-    /**
-     * Delete every selected row under one undo transaction (UXP
-     * `onDeleteCmd`, which loops the whole `getSelectedNodeList()`).
-     */
-    const onRemoveRow = useCallback(() => {
-        const t = requireTarget()
-        if (!t || !cm || selectedRows.size === 0) return
-        fireService(cm, 'removePaintEntries', { ...t, idxs: [...selectedRows] })
-        setSelectedRow(null)
-    }, [cm, requireTarget, selectedRows, setSelectedRow])
-
-    const onMoveRow = useCallback(
-        (dir: 'up' | 'down') => {
-            const t = requireTarget()
-            if (!t || !cm || selectedRow === null) return
-            const toIdx = dir === 'up' ? selectedRow - 1 : selectedRow + 1
-            if (toIdx < 0 || toIdx >= entries.length) return
-            fireService(cm, 'movePaintEntry', {
-                ...t,
-                fromIdx: selectedRow,
-                toIdx,
-            })
-            setSelectedRow(toIdx)
-        },
-        [cm, requireTarget, selectedRow, entries.length, setSelectedRow],
-    )
-
-    const onRemoveAllRows = useCallback(() => {
-        const t = requireTarget()
-        if (!t || !cm) return
-        fireService(cm, 'clearPaintEntries', t)
-        setSelectedRow(null)
-    }, [cm, requireTarget, setSelectedRow])
-
-    /**
-     * Copy / Cut the selected row onto the OS clipboard.
-     *
-     * Cut deletes in the worker first and writes afterwards, mirroring UXP
-     * `onCut` (= onCopy + onDeleteCmd) as one undo step. If the clipboard
-     * write then fails the rows are gone, but a single Undo restores them.
-     */
-    const onClipboardTake = useCallback(
-        (mode: 'copy' | 'cut') => {
-            const t = requireTarget()
-            if (!t || !cm || selectedRows.size === 0) return
-            const name = mode === 'cut' ? 'cutPaintEntries' : 'copyPaintEntries'
-            cm.invokeService(name, { ...t, idxs: [...selectedRows] })
-                .then(async (res) => {
-                    if (!res.ok || res.entries.length === 0) return
-                    if (mode === 'cut') setSelectedRow(null)
-                    const w = await window.electronAPI?.invoke(
-                        IPC.CLIPBOARD_CUEMOL_WRITE,
-                        { kind: 'paint', entries: res.entries },
-                    )
-                    setCanPastePaint(w?.ok === true)
-                })
-                .catch((err: unknown) => console.warn(`${name} failed:`, err))
-        },
-        [cm, requireTarget, selectedRows, setSelectedRow],
-    )
-
-    const onPasteRows = useCallback(() => {
-        const t = requireTarget()
-        if (!t || !cm) return
-        void (async () => {
-            try {
-                const clip = await window.electronAPI?.invoke(
-                    IPC.CLIPBOARD_CUEMOL_READ,
-                )
-                if (clip?.kind !== 'paint') {
-                    setCanPastePaint(false)
-                    return
-                }
-                // UXP `_getPaintSelImpl`: insert before the *first* selected
-                // row (`elems[0]`), or append when nothing is selected.
-                const anchor =
-                    selectedRows.size === 0 ? null : Math.min(...selectedRows)
-                const res = await cm.invokeService('pastePaintEntries', {
-                    ...t,
-                    idx: anchor,
-                    entries: clip.entries,
-                })
-                if (res.ok) setSelectedRow(res.startIdx)
-            } catch (err) {
-                console.warn('pastePaintEntries failed:', err)
-            }
-        })()
-    }, [cm, requireTarget, selectedRows, setSelectedRow])
-
-    const onUpdateCell = useCallback(
-        (idx: number, field: 'selStr' | 'colorValue', value: string) => {
-            const t = requireTarget()
-            const cur = entries[idx]
-            if (!t || !cm || !cur) return
-            fireService(cm, 'updatePaintEntry', {
-                ...t,
-                idx,
-                selStr: field === 'selStr' ? value : cur.selStr,
-                colorValue: field === 'colorValue' ? value : cur.colorValue,
-            })
-        },
-        [cm, requireTarget, entries],
-    )
-
-    const onDefaultColorCommit = useCallback(
-        (color: string) => {
-            const t = requireTarget()
-            if (!t || !cm) return
-            fireService(cm, 'setRendererDefaultColor', {
-                ...t,
-                colorValue: color,
-            })
-        },
-        [cm, requireTarget],
-    )
-
-    /**
-     * Generic ColoringScheme property commit (CPK col_X, Rainbow start_hue,
-     * Bfac mode, etc.). Routes through `setColoringProp` which mirrors
-     * UXP `commitPropChange` (materialize on default + setProp under undo).
-     */
-    const onSetColoringProp = useCallback(
-        (propName: string, value: string | number) => {
-            const t = requireTarget()
-            if (!t || !cm) return
-            fireService(cm, 'setColoringProp', {
-                ...t,
-                propName,
-                propValue: value,
-            })
-        },
-        [cm, requireTarget],
-    )
-
-    /**
-     * Elepot widgets commit directly on the renderer (lowpar/midpar/highpar
-     * etc.) -- not on a ColoringScheme -- so they route through a separate
-     * service. Mirrors UXP `commitElepotPropChange`.
-     */
-    const onSetElepotProp = useCallback(
-        (propName: string, value: string | number | boolean) => {
-            const t = requireTarget()
-            if (!t || !cm) return
-            fireService(cm, 'setRendererElepotProp', {
-                ...t,
-                propName,
-                propValue: value,
-            })
-        },
-        [cm, requireTarget],
-    )
-
-    /**
-     * "Coloring mol" selector commit: write the MOLFANC reference-molecule
-     * name into the renderer's `target` property.
-     */
-    const onSetColoringTarget = useCallback(
-        (targetName: string) => {
-            const t = requireTarget()
-            if (!t || !cm) return
-            fireService(cm, 'setRendererColoringTarget', {
-                ...t,
-                targetName,
-            })
-        },
-        [cm, requireTarget],
-    )
+    const {
+        onSelectMode, onAddRow, onRemoveRow, onMoveRow, onRemoveAllRows,
+        onClipboardTake, onPasteRows, onUpdateCell, onDefaultColorCommit,
+        onSetColoringProp, onSetElepotProp, onSetColoringTarget,
+    } = useColorPaneActions({
+        cm, requireTarget, entries, selectedRow, selectedRows, setSelectedRow,
+        setCanPastePaint,
+    })
 
     // -- Deck routing --
     const renderDeck = (): React.ReactNode => {
